+++ packages/pudu/v0.1/src/Pudu/Semantic/Prelude.hs
@@ -86,7 +86,7 @@ effectValueNames =
   , "mutexOpen", "mutexAcquire", "mutexRelease"
   , "cellOpen", "cellGet", "cellSwap"
   , "secureRandomBytes"
-  , "desktopOpen", "desktopPresent", "desktopPump", "desktopInputs", "desktopClose", "audioDevicePlay"
+  , "desktopOpen", "desktopPresent", "desktopPump", "desktopInputs", "desktopClipboardRead", "desktopClipboardWrite", "desktopClose", "audioDevicePlay"
   , "audioStreamOpen", "audioStreamWrite", "audioStreamPause", "audioStreamResume"
   , "audioStreamVolume", "audioStreamSnapshot", "audioStreamClose"
   , "arguments", "environment", "temporaryPath", "userHome"